+++ src/models/particle.rs
@@ -1,5 +1,4 @@
-use super::Vector;
-use geometry::Advance;
+use geometry::{Advance, Vector};
 
 /// A model representing a particle
 ///
@@ -8,19 +7,19 @@ use geometry::Advance;
 /// player or an enemy is killed
 pub struct Particle {
     pub vector: Vector,
-    pub ttl: f64
+    pub ttl: f32
 }
 
 derive_position_direction!(Particle);
 
 impl Particle {
     /// Create a particle with the given vector and time to live in seconds
-    pub fn new(vector: Vector, ttl: f64) -> Particle {
+    pub fn new(vector: Vector, ttl: f32) -> Particle {
         Particle { vector: vector, ttl: ttl }
     }
 
     /// Update the particle
-    pub fn update(&mut self, elapsed_time: f64) {
+    pub fn update(&mut self, elapsed_time: f32) {
         self.ttl -= elapsed_time;
         let speed = 500.0 * self.ttl * self.ttl;
         self.advance(elapsed_time * speed);
